Replaced SyncAdapter with JobService. Currency is now a universe defined by the backend. Autoselecting of account currency on add record screen. Autoconversion to account currency the amount specified when creating account if the currency selected does not match the account currency. All amounts are stored in cents and shown in 100 cents units.

+++ app/src/main/java/nacholab/showmethemoney/model/MoneyAccount.java
@@ -16,7 +16,7 @@ public class MoneyAccount extends DBModel {
 
     @Expose
     @Column
-    private float balance;
+    private int balance;
 
     private Currency currencyObject;
 
@@ -47,11 +47,11 @@ public class MoneyAccount extends DBModel {
         this.currency = currency;
     }
 
-    public float getBalance() {
+    public int getBalance() {
         return balance;
     }
 
-    public void setBalance(float balance) {
+    public void setBalance(int balance) {
         this.balance = balance;
     }
